+++ docs/zh-cn/user-guide/CHANGELOG.md
@@ -1,5 +1,10 @@
 # 特性更新日志
 
+## next version
+
+- feat(core/ui/quick-open): 支持类 vscode 的 input-box & quick-pick
+- fix(core/ui/ribbon): 修复在 macOS 下不能打开 APP 设置地问题 ([#36](https://github.com/typora-community-plugin/typora-community-plugin/issues/36))
+
 ## v2.3.30
 
 - fix(core/ui/component/notice): 修复不能在 setMessage 之前 setCloseable 的问题